+++ HOTELATENDINA/.idea/artifacts/HOTELATENDINA_jar.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="HOTELATENDINA:jar">
+    <output-path>$PROJECT_DIR$/out/artifacts/HOTELATENDINA_jar</output-path>
+    <root id="archive" name="HOTELATENDINA.jar">
+      <element id="module-output" name="HOTELATENDINA" />
+    </root>
+  </artifact>
+</component>
