--- HOTELATENDINA/HOTEL.form
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.ui.HOTELATENDINA.HOTEL">
-  <grid id="27dc6" binding="Panel1" layout-manager="GridLayoutManager" row-count="8" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="500" height="400"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <component id="6244c" class="javax.swing.JButton" binding="Pulsantediconferma">
-        <constraints>
-          <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Conferma"/>
-        </properties>
-      </component>
-      <component id="4c277" class="javax.swing.JButton" binding="InfoPulsante">
-        <constraints>
-          <grid row="5" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Informazioni"/>
-        </properties>
-      </component>
-      <component id="4be5b" class="javax.swing.JButton" binding="PulsanteVisualizzaFoto">
-        <constraints>
-          <grid row="6" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Visualizza Foto"/>
-        </properties>
-      </component>
-      <component id="252a6" class="javax.swing.JButton" binding="PulsanteEsci">
-        <constraints>
-          <grid row="7" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Esci"/>
-        </properties>
-      </component>
-      <component id="306b5" class="javax.swing.JCheckBox" binding="FumatoriCheckBox1" custom-create="true">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <enabled value="false"/>
-          <text value="Fumatori"/>
-        </properties>
-      </component>
-      <component id="a7ce7" class="javax.swing.JCheckBox" binding="CameraMatrimonialeCheckBox2" custom-create="true">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <enabled value="false"/>
-          <text value="Camera Matrimoniale"/>
-        </properties>
-      </component>
-      <component id="17cc9" class="javax.swing.JCheckBox" binding="VistamareCheckBox3" custom-create="true">
-        <constraints>
-          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <enabled value="false"/>
-          <text value="Vista Mare"/>
-        </properties>
-      </component>
-      <component id="5fe9c" class="javax.swing.JCheckBox" binding="ParcheggioCheckBox4" custom-create="true">
-        <constraints>
-          <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <enabled value="false"/>
-          <selected value="false"/>
-          <text value="Parcheggio"/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
