+++ HOTELATENDINA/src/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: com.ui.HOTELATENDINA.App
+
