+++ HOTELATENDINA/src/com/ui/HOTELATENDINA/App.form
@@ -1,0 +1,100 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.ui.HOTELATENDINA.App">
+  <grid id="27dc6" binding="panel1" layout-manager="GridLayoutManager" row-count="9" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+    <margin top="0" left="0" bottom="0" right="0"/>
+    <constraints>
+      <xy x="20" y="20" width="500" height="400"/>
+    </constraints>
+    <properties/>
+    <border type="none"/>
+    <children>
+      <component id="5fdbe" class="javax.swing.JCheckBox" binding="FumatoriCheckBox1">
+        <constraints>
+          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text value="Fumatori"/>
+        </properties>
+      </component>
+      <component id="3bfa" class="javax.swing.JCheckBox" binding="CameraMatrimonialeCheckBox2">
+        <constraints>
+          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="23"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Camera Matrimoniale"/>
+        </properties>
+      </component>
+      <component id="88a39" class="javax.swing.JCheckBox" binding="VistamareCheckBox3">
+        <constraints>
+          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="23"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Vista mare"/>
+        </properties>
+      </component>
+      <component id="182f8" class="javax.swing.JCheckBox" binding="ParcheggioCheckBox4">
+        <constraints>
+          <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="23"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Parcheggio"/>
+        </properties>
+      </component>
+      <component id="68077" class="javax.swing.JButton" binding="Pulsantediconferma">
+        <constraints>
+          <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="30"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Conferma"/>
+        </properties>
+      </component>
+      <component id="a0588" class="javax.swing.JButton" binding="InfoPulsante">
+        <constraints>
+          <grid row="5" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="30"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Info"/>
+        </properties>
+      </component>
+      <component id="f8a0e" class="javax.swing.JButton" binding="PulsanteVisualizzaFoto">
+        <constraints>
+          <grid row="6" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="30"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Visualizza Foto"/>
+        </properties>
+      </component>
+      <component id="1f1d9" class="javax.swing.JButton" binding="PulsanteEsci">
+        <constraints>
+          <grid row="8" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false">
+            <preferred-size width="513" height="30"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value="Esci"/>
+        </properties>
+      </component>
+      <component id="2750" class="javax.swing.JLabel" binding="HotelFoto" custom-create="true">
+        <constraints>
+          <grid row="7" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text value=""/>
+          <visible value="false"/>
+        </properties>
+      </component>
+    </children>
+  </grid>
+</form>
